+++ crate/model/src/info_graph.rs
@@ -12,45 +12,111 @@ use crate::{
     theme::Theme,
 };
 
-pub use self::{graph_dir::GraphDir, info_graph_builder::InfoGraphBuilder, tag::Tag};
+pub use self::{graph_dir::GraphDir, tag::Tag};
 
 mod graph_dir;
-mod info_graph_builder;
 mod tag;
 
 #[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize)]
 #[serde(default)]
 pub struct InfoGraph {
     /// Direction of the graph, `vertical` or `horizontal`.
-    pub(crate) direction: GraphDir,
+    pub direction: GraphDir,
     /// Nested nodes.
-    pub(crate) hierarchy: NodeHierarchy,
+    pub hierarchy: NodeHierarchy,
     /// Each node's name.
-    pub(crate) node_names: NodeNames,
+    pub node_names: NodeNames,
     /// Each node's description.
-    pub(crate) node_descs: NodeDescs,
+    pub node_descs: NodeDescs,
     /// Each node's emoji.
-    pub(crate) node_emojis: NodeEmojis,
+    pub node_emojis: NodeEmojis,
     /// Tags associated with each node.
-    pub(crate) node_tags: NodeTags,
+    pub node_tags: NodeTags,
     /// Logical / ordering dependencies.
-    pub(crate) edges: Edges,
+    pub edges: Edges,
     /// Each edge's description.
-    pub(crate) edge_descs: EdgeDescs,
+    pub edge_descs: EdgeDescs,
     /// Tags to associate with nodes.
-    pub(crate) tags: IndexMap<TagId, Tag>,
+    pub tags: IndexMap<TagId, Tag>,
     /// Additional attributes specifically for GraphViz.
-    pub(crate) graphviz_attrs: GraphvizAttrs,
+    pub graphviz_attrs: GraphvizAttrs,
     /// Theme that controls the CSS classes to add to elements.
-    pub(crate) theme: Theme,
+    pub theme: Theme,
     /// Additional CSS to add in the spreadsheet.
-    pub(crate) css: String,
+    pub css: String,
 }
 
 impl InfoGraph {
-    /// Returns a builder to instantiate an `InfoGraph`.
-    pub fn builder() -> InfoGraphBuilder {
-        InfoGraphBuilder::default()
+    /// Sets the direction of the graph, `vertical` or `horizontal`.
+    pub fn with_direction(mut self, direction: GraphDir) -> Self {
+        self.direction = direction;
+        self
+    }
+
+    /// Sets the nested nodes.
+    pub fn with_hierarchy(mut self, hierarchy: NodeHierarchy) -> Self {
+        self.hierarchy = hierarchy;
+        self
+    }
+
+    /// Sets the map of node names.
+    pub fn with_node_names(mut self, node_names: NodeNames) -> Self {
+        self.node_names = node_names;
+        self
+    }
+
+    /// Sets the map of node descriptions.
+    pub fn with_node_descs(mut self, node_descs: NodeDescs) -> Self {
+        self.node_descs = node_descs;
+        self
+    }
+
+    /// Sets the map of node emojis.
+    pub fn with_node_emojis(mut self, node_emojis: NodeEmojis) -> Self {
+        self.node_emojis = node_emojis;
+        self
+    }
+
+    /// Sets the tags associated with each node.
+    pub fn with_node_tags(mut self, node_tags: NodeTags) -> Self {
+        self.node_tags = node_tags;
+        self
+    }
+
+    /// Sets the logical / ordering dependencies.
+    pub fn with_edges(mut self, edges: Edges) -> Self {
+        self.edges = edges;
+        self
+    }
+
+    /// Sets the map of edge descriptions.
+    pub fn with_edge_descs(mut self, edge_descs: EdgeDescs) -> Self {
+        self.edge_descs = edge_descs;
+        self
+    }
+
+    /// Sets the tags to associate with nodes.
+    pub fn with_tags(mut self, tags: IndexMap<TagId, Tag>) -> Self {
+        self.tags = tags;
+        self
+    }
+
+    /// Sets the additional attributes specifically for GraphViz.
+    pub fn with_graphviz_attrs(mut self, graphviz_attrs: GraphvizAttrs) -> Self {
+        self.graphviz_attrs = graphviz_attrs;
+        self
+    }
+
+    /// Sets the theme that controls the CSS classes to add to elements.
+    pub fn with_theme(mut self, theme: Theme) -> Self {
+        self.theme = theme;
+        self
+    }
+
+    /// Sets the additional CSS to add in the spreadsheet.
+    pub fn with_css(mut self, css: String) -> Self {
+        self.css = css;
+        self
     }
 
     /// Returns the direction of the graph, `vertical` or `horizontal`.